Refactor command test packages
