+++ src/test/java/seedu/address/logic/commands/addressbookcommands/ExitCommandTest.java
@@ -1,10 +1,11 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.addressbookcommands;
 
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.address.logic.commands.addressbookcommands.ExitCommand.MESSAGE_EXIT_ACKNOWLEDGEMENT;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.addressbookcommands.ExitCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
