+++ src/test/java/seedu/address/logic/commands/addressbookcommands/HelpCommandTest.java
@@ -1,10 +1,11 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.addressbookcommands;
 
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.address.logic.commands.addressbookcommands.HelpCommand.SHOWING_HELP_MESSAGE;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.addressbookcommands.HelpCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
